backend/requirements.txt
pandas==2.2.0
mysql-connector-python==8.3.0
openpyxl==3.1.2
flask==3.0.0
matplotlib==3.8.2